requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.49.0
django-oidc-provider==0.7.0
future==0.18.2
idna==3.3
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodomex==3.14.1
pyjwkest==1.4.2
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
